+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 * nostr: remove `bech32` from the public API ([Yuki Kishimoto])
 * nostr: remove `Keys::from_public_key` ([Yuki Kishimoto])
 * nostr: remove `tracing` dep ([Yuki Kishimoto])
+* nostr: remove impl `fmt::Display` for `SecretKey` ([Yuki Kishimoto])
 * js(nostr): remove `Keys::vanity` ([Yuki Kishimoto])
 
 ## [v0.34.0]